# packages/contracts/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "JB API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/v1/portfolio": {
   "get": {
    "tags": [
     "portfolio"
    ],
    "summary": "Get Portfolio",
    "operationId": "get_portfolio_api_v1_portfolio_get",
    "parameters": [
     {
      "name": "x-demo-user",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Demo-User"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Portfolio"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/portfolio/profile/parse": {
   "post": {
    "tags": [
     "portfolio"
    ],
    "summary": "Parse Profile",
    "description": "Free text about yourself and your goals -> RiskProfile (Grok structured). Stored for suggest().",
    "operationId": "parse_profile_api_v1_portfolio_profile_parse_post",
    "parameters": [
     {
      "name": "x-demo-user",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Demo-User"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProfileIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/portfolio/profile": {
   "get": {
    "tags": [
     "portfolio"
    ],
    "summary": "Get Profile",
    "operationId": "get_profile_api_v1_portfolio_profile_get",
    "parameters": [
     {
      "name": "x-demo-user",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Demo-User"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/portfolio/suggest": {
   "post": {
    "tags": [
     "portfolio"
    ],
    "summary": "Suggest",
    "description": "Half Kelly over every market (or the user's filters). value = user's own value if given,\nelse the model value. Role D layers vector matching and Grok narrative on top of this.",
    "operationId": "suggest_api_v1_portfolio_suggest_post",
    "parameters": [
     {
      "name": "x-demo-user",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Demo-User"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/SuggestIn"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Suggestion"
         },
         "title": "Response Suggest Api V1 Portfolio Suggest Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/surveillance/audit": {
   "get": {
    "tags": [
     "surveillance"
    ],
    "summary": "Audit",
    "operationId": "audit_api_v1_surveillance_audit_get",
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 200,
       "title": "Limit"
      }
     },
     {
      "name": "actor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Actor"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/surveillance/flags": {
   "get": {
    "tags": [
     "surveillance"
    ],
    "summary": "Flags",
    "description": "Rules flags, each reviewed by Grok and K2 when keys are set (cached per flag). Disagreement marks `disputed`.",
    "operationId": "flags_api_v1_surveillance_flags_get",
    "parameters": [
     {
      "name": "market_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Market Id"
      }
     },
     {
      "name": "review",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": true,
       "title": "Review"
      }
     },
     {
      "name": "max_reviews",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 4,
       "title": "Max Reviews"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Flag"
         },
         "title": "Response Flags Api V1 Surveillance Flags Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/surveillance/redteam": {
   "post": {
    "tags": [
     "surveillance"
    ],
    "summary": "Red Team",
    "description": "Grok plays a manipulator against one market; the surveillance layer has to catch it.\nReturns the attack plan and the orders it placed. Flags appear over the next rounds.",
    "operationId": "red_team_api_v1_surveillance_redteam_post",
    "parameters": [
     {
      "name": "x-demo-user",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Demo-User"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/RedTeamIn"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "surveillance"
    ],
    "summary": "Red Team Log",
    "operationId": "red_team_log_api_v1_surveillance_redteam_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Flag": {
    "properties": {
     "_id": {
      "type": "string",
      "title": "Id"
     },
     "market_id": {
      "type": "string",
      "title": "Market Id"
     },
     "batch_id": {
      "type": "string",
      "title": "Batch Id"
     },
     "rule": {
      "type": "string",
      "title": "Rule"
     },
     "severity": {
      "type": "string",
      "enum": [
       "benign",
       "low",
       "medium",
       "high"
      ],
      "title": "Severity"
     },
     "subjects": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Subjects"
     },
     "explanation": {
      "type": "string",
      "title": "Explanation"
     },
     "reviewer": {
      "type": "string",
      "enum": [
       "rules",
       "grok",
       "k2"
      ],
      "title": "Reviewer"
     },
     "reviews": {
      "items": {
       "$ref": "#/components/schemas/Review"
      },
      "type": "array",
      "title": "Reviews"
     },
     "disputed": {
      "type": "boolean",
      "title": "Disputed"
     },
     "t": {
      "type": "string",
      "title": "T"
     }
    },
    "type": "object",
    "required": [
     "_id",
     "market_id",
     "batch_id",
     "rule",
     "severity",
     "subjects",
     "explanation",
     "reviewer",
     "reviews",
     "disputed",
     "t"
    ],
    "title": "Flag"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "Pnl": {
    "properties": {
     "realized": {
      "type": "number",
      "title": "Realized"
     },
     "unrealized": {
      "type": "number",
      "title": "Unrealized"
     },
     "total": {
      "type": "number",
      "title": "Total"
     }
    },
    "type": "object",
    "required": [
     "realized",
     "unrealized",
     "total"
    ],
    "title": "Pnl"
   },
   "Portfolio": {
    "properties": {
     "user_id": {
      "type": "string",
      "title": "User Id"
     },
     "cash": {
      "type": "number",
      "title": "Cash"
     },
     "reserved_cash": {
      "type": "number",
      "title": "Reserved Cash"
     },
     "equity": {
      "type": "number",
      "title": "Equity"
     },
     "pnl": {
      "$ref": "#/components/schemas/Pnl"
     },
     "positions": {
      "items": {
       "$ref": "#/components/schemas/Position"
      },
      "type": "array",
      "title": "Positions"
     }
    },
    "type": "object",
    "required": [
     "user_id",
     "cash",
     "reserved_cash",
     "equity",
     "pnl",
     "positions"
    ],
    "title": "Portfolio"
   },
   "Position": {
    "properties": {
     "market_id": {
      "type": "string",
      "title": "Market Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "qty": {
      "type": "number",
      "title": "Qty"
     },
     "avg_cost": {
      "type": "number",
      "title": "Avg Cost"
     },
     "last": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last"
     },
     "value": {
      "type": "number",
      "title": "Value"
     },
     "pnl": {
      "type": "number",
      "title": "Pnl"
     }
    },
    "type": "object",
    "required": [
     "market_id",
     "name",
     "qty",
     "avg_cost",
     "last",
     "value",
     "pnl"
    ],
    "title": "Position"
   },
   "ProfileIn": {
    "properties": {
     "text": {
      "type": "string",
      "title": "Text"
     }
    },
    "type": "object",
    "required": [
     "text"
    ],
    "title": "ProfileIn"
   },
   "RedTeamIn": {
    "properties": {
     "market_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Market Id"
     }
    },
    "type": "object",
    "title": "RedTeamIn"
   },
   "Review": {
    "properties": {
     "reviewer": {
      "type": "string",
      "enum": [
       "rules",
       "grok",
       "k2"
      ],
      "title": "Reviewer"
     },
     "severity": {
      "type": "string",
      "enum": [
       "benign",
       "low",
       "medium",
       "high"
      ],
      "title": "Severity"
     }
    },
    "type": "object",
    "required": [
     "reviewer",
     "severity"
    ],
    "title": "Review"
   },
...
   "SuggestIn": {
    "properties": {
     "bankroll": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bankroll"
     },
     "kelly_multiplier": {
      "type": "number",
      "title": "Kelly Multiplier",
      "default": 0.5
     },
     "own_values": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Own Values"
     },
     "states": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "States"
     },
     "categories": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Categories"
     },
     "exclude_held": {
      "type": "boolean",
      "title": "Exclude Held",
      "default": true
     }
    },
    "type": "object",
    "title": "SuggestIn"
   },
   "Suggestion": {
    "properties": {
     "company": {
      "$ref": "#/components/schemas/SuggestionCompany"
     },
     "price": {
      "type": "number",
      "title": "Price"
     },
     "model_value": {
      "type": "number",
      "title": "Model Value"
     },
     "edge": {
      "type": "number",
      "title": "Edge"
     },
     "sigma": {
      "type": "number",
      "title": "Sigma"
     },
     "kelly_fraction": {
      "type": "number",
      "title": "Kelly Fraction"
     },
     "suggested_usd": {
      "type": "number",
      "title": "Suggested Usd"
     },
     "why": {
      "type": "string",
      "title": "Why"
     }
    },
    "type": "object",
    "required": [
     "company",
     "price",
     "model_value",
     "edge",
     "sigma",
     "kelly_fraction",
     "suggested_usd",
     "why"
    ],
    "title": "Suggestion"
   },
   "SuggestionCompany": {
    "properties": {
     "_id": {
      "type": "string",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "city": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "City"
     },
     "state": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "State"
     },
     "category": {
      "type": "string",
      "title": "Category"
     }
    },
    "type": "object",
    "required": [
     "_id",
     "name",
     "city",
     "state",
     "category"
    ],
    "title": "SuggestionCompany"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}